requirements.txt
flask
google-generativeai
Pillow
Werkzeug
